Link Author's API calls to authenticated User

+++ src/registry/api.py
@@ -3,9 +3,17 @@ from typing import List, Optional
 from django.db.models import Q
 
 from .models import Author, Model, Prediction
+from main.models import CustomUser
 from ninja import Router
 from ninja.orm.fields import AnyObject
-from .schema import AuthorSchema, ModelSchema, NotFoundSchema, PredictionSchema, Schema
+from .schema import (
+    AuthorSchema,
+    ModelSchema,
+    NotFoundSchema,
+    PredictionSchema,
+    ForbiddenSchema,
+    Schema,
+)
 
 router = Router()
 
@@ -15,6 +23,12 @@ class AuthorIn(Schema):
     """Input for the request's body"""
 
     user: str
+    institution: str
+
+
+class AuthorInPost(Schema):
+    """Input for POST's request body"""
+
     institution: str
 
 
@@ -31,43 +45,66 @@ def list_authors(request, name: Optional[str] = None):
     return authors
 
 
-@router.get("/authors/{author_id}", response={200: AuthorSchema, 404: NotFoundSchema})
-def get_author(request, author_id: int):
-    """Gets author by id"""
+@router.get("/authors/{username}", response={200: AuthorSchema, 404: NotFoundSchema})
+def get_author(request, username: str):
+    """Gets author by Github username"""
     try:
-        author = Author.objects.get(pk=author_id)
-        return (200, author)
+        author = Author.objects.get(user__username=username)
+        return 200, author
     except Author.DoesNotExist:
-        return (404, {"message": "Author not found"})
+        return 404, {"message": "Author not found"}
 
 
-@router.post("/authors/", response={201: AuthorSchema})
+@router.post(
+    "/authors/", response={201: AuthorSchema, 404: NotFoundSchema, 403: ForbiddenSchema}
+)
 def create_author(request, payload: AuthorIn):
-    """Posts author to database"""
-    author = Author.objects.create(**payload.dict())
-    return (201, author)
-
-
-@router.put("/authors/{author_id}", response={201: AuthorSchema, 404: NotFoundSchema})
-def update_author(request, author_id: int, payload: AuthorIn):
-    """Updates author"""
+    """Posts author to database, requires a CustomUser to be created"""
     try:
-        author = Author.objects.get(pk=author_id)
+        user = CustomUser.objects.get(username=payload.user)
+        try:
+            author = Author.objects.get(user__username=payload.user)
+            if author:
+                return 403, {"message": f"Author '{author}' already registered"}
+        except Author.DoesNotExist:
+            author = Author.objects.create(user=user, institution=payload.institution)
+            return 201, author
+    except CustomUser.DoesNotExist:
+        return 404, {"message": f"User '{payload.user}' does not exist."}
 
-        for attr, value in payload.dict().items():
-            setattr(author, attr, value)
 
+@router.put(
+    "/authors/{username}",
+    response={201: AuthorSchema, 403: ForbiddenSchema, 404: NotFoundSchema},
+)
+def update_author(request, username: str, payload: AuthorInPost):
+    """
+    Updates author. It is not possible to change Author's
+    user and this post method can only be called by the user
+    """
+    try:
+        author = Author.objects.get(user__username=username)
+
+        if request.user != author.user:  # TODO: Enable admins here
+            return 403, {"message": "You are not authorized to update this author."}
+
+        author.institution = payload.institution
         author.save()
-        return (201, author)
+        return 201, author
     except Author.DoesNotExist:
-        return (404, {"message": "Author not found"})
+        return 404, {"message": "Author not found"}
 
 
-@router.delete("/authors/{author_id}", response={204: None, 404: NotFoundSchema})
-def delete_author(request, author_id: int):
+@router.delete(
+    "/authors/{username}",
+    response={204: None, 403: ForbiddenSchema, 404: NotFoundSchema},
+)
+def delete_author(request, username: str):
     """Deletes author"""
     try:
-        author = Author.objects.get(pk=author_id)
+        author = Author.objects.get(user__username=username)
+        if request.user != author.user:  # TODO: Enable admins here
+            return 403, {"message": "You are not authorized to delete this author."}
         author.delete()
         return 204
     except Author.DoesNotExist:
